requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-docx>=1.1.0
openai>=1.54.3
anthropic>=0.40.0
httpx>=0.27.2
tiktoken>=0.8.0
streamlit>=1.31.0
